backend/requirements.txt
# Core web framework
fastapi==0.115.13
uvicorn==0.34.3

# Database
sqlalchemy==1.4.42
psycopg2-binary==2.9.10

# AI and document processing
openai==1.90.0
PyMuPDF==1.24.0
google-cloud-storage==2.14.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.0
pydantic==2.8.2
